feat(catalog): fixed packages with whole-number quantities, Persian digits
Replace the per-kilogram decimal-quantity model with fixed packages:
boxed sweets come as their own variants (جعبه نیم/یک/دو کیلویی، هر جعبه
با قیمت خودش) and jar products are ۴۵۰-gram containers — quantity is
always a whole number of packages.

- drop qty_step/is_weighted; min_order_qty, stock_qty and order-item
  quantity become integers (data migrations normalize legacy rows;
  line_total snapshots stay untouched)
- parse_quantity() accepts Persian/Arabic digits, rejects fractions
  with a Persian error; carts drop legacy fractional session lines
- qty steppers are text inputs rendered in Persian digits (±1 steps);
  remove the kg template filter, quantities display via |fa
- product subtitle field for secondary traits, shown smaller under the
  name (حلوا زرده اعلا ← روغن حیوانی); rename انگشت‌پیچ اعلا to ویژه
- reseed: per-box SKUs with derived prices, retire old per-kg SKUs and
  variants of retired products (deactivate, never delete)

+++ apps/orders/models.py
@@ -120,7 +120,7 @@ class OrderItem(models.Model):
     product_name = models.CharField(max_length=200)
     variant_label = models.CharField(max_length=100, blank=True)
     unit_label = models.CharField(max_length=20)
-    quantity = models.DecimalField(max_digits=8, decimal_places=2)
+    quantity = models.PositiveIntegerField()
     unit_price = models.DecimalField(max_digits=12, decimal_places=0)
     line_total = models.DecimalField(max_digits=14, decimal_places=0)
